requirements.txt
django-tastypie==0.9.11